+++ src/lib/modules/api/models/percentile-historic-indicator-query-params.model.ts
@@ -1,0 +1,6 @@
+import { IndicatorQueryParams } from './indicator-query-params.model';
+
+export interface PercentileHistoricIndicatorQueryParams extends IndicatorQueryParams {
+  historic_range: Number;
+  percentile: Number;
+}